# mobile/front/sip_log_mobile/API_Contracts/api-docsV11.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://ec2-52-2-112-100.compute-1.amazonaws.com:8081","description":"Generated server url"}],"paths":{"/api/v1/usuarios/me":{"get":{"tags":["usuario-bff-controller"],"operationId":"getMeuPerfil","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PerfilDTO"}}}}}},"put":{"tags":["usuario-bff-controller"],"operationId":"atualizarMeuPerfil","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["usuario-bff-controller"],"operationId":"deletarMinhaConta","responses":{"200":{"description":"OK"}}}},"/api/v1/experiencias/{id}":{"get":{"tags":["experiencia-bff-controller"],"operationId":"buscarExperienciaPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegistroExperienciaDTO"}}}}}},"put":{"tags":["experiencia-bff-controller"],"operationId":"editarExperiencia","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NovaExperienciaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegistroExperienciaDTO"}}}}}},"delete":{"tags":["experiencia-bff-controller"],"operationId":"deletarExperiencia","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/experiencias/{id}/comentarios/{idComentario}":{"put":{"tags":["experiencia-bff-controller"],"operationId":"editarComentario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idComentario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NovoComentarioDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ComentarioResponseDTO"}}}}}},"delete":{"tags":["experiencia-bff-controller"],"operationId":"deletarComentario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idComentario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/usuarios/{idAlvo}/seguir":{"post":{"tags":["usuario-bff-controller"],"operationId":"alternarSeguir","parameters":[{"name":"idAlvo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/usuarios/sync":{"post":{"tags":["usuario-bff-controller"],"operationId":"sincronizarUsuario","responses":{"200":{"description":"OK"}}}},"/api/v1/upload":{"post":{"tags":["upload-bff-controller"],"operationId":"uploadFoto","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UploadResponseDTO"}}}}}}},"/api/v1/experiencias":{"post":{"tags":["experiencia-bff-controller"],"operationId":"registrarNovaExperiencia","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NovaExperienciaDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/experiencias/{id}/curtir":{"post":{"tags":["experiencia-bff-controller"],"operationId":"alternarCurtida","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/experiencias/{id}/comentarios":{"get":{"tags":["experiencia-bff-controller"],"operationId":"listarComentarios","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveComentarioResponseDTO"}}}}}},"post":{"tags":["experiencia-bff-controller"],"operationId":"adicionarComentario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NovoComentarioDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ComentarioResponseDTO"}}}}}}},"/api/v1/bebidas":{"post":{"tags":["bebida-bff-controller"],"operationId":"adicionarBebida","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NovaBebidaDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BebidaResumoDTO"}}}}}}},"/api/v1/notificacoes/{id}/lida":{"patch":{"tags":["notificacao-bff-controller"],"operationId":"marcarComoLida","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/usuarios/{idUsuario}":{"get":{"tags":["usuario-bff-controller"],"operationId":"getPerfilUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PerfilDTO"}}}}}}},"/api/v1/usuarios/{idUsuario}/seguindo":{"get":{"tags":["usuario-bff-controller"],"operationId":"obterSeguindo","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveUsuarioResumoDTO"}}}}}}},"/api/v1/usuarios/{idUsuario}/seguidores":{"get":{"tags":["usuario-bff-controller"],"operationId":"getSeguidores","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveUsuarioResumoDTO"}}}}}}},"/api/v1/usuarios/buscar":{"get":{"tags":["usuario-bff-controller"],"operationId":"buscarUsuarios","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}},{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveUsuarioResumoDTO"}}}}}}},"/api/v1/notificacoes":{"get":{"tags":["notificacao-bff-controller"],"operationId":"listarNotificacoes","parameters":[{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveNotificacaoResponseDTO"}}}}}}},"/api/v1/notificacoes/nao-lidas/count":{"get":{"tags":["notificacao-bff-controller"],"operationId":"contarNaoLidas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContagemNotificacoesDTO"}}}}}}},"/api/v1/feed/usuarios/{idUsuario}":{"get":{"tags":["feed-bff-controller"],"operationId":"getFeedDeUsuario","parameters":[{"name":"idUsuario","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveFeedResponseDTO"}}}}}}},"/api/v1/feed/me":{"get":{"tags":["feed-bff-controller"],"operationId":"getFeedMe","parameters":[{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveFeedResponseDTO"}}}}}}},"/api/v1/feed/global":{"get":{"tags":["feed-bff-controller"],"operationId":"getFeedGlobal","parameters":[{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveFeedResponseDTO"}}}}}}},"/api/v1/feed/amigos":{"get":{"tags":["feed-bff-controller"],"operationId":"getFeedAmigos","parameters":[{"name":"pagina","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaBffDTOReciveFeedResponseDTO"}}}}}}},"/api/v1/bebidas/{id}":{"get":{"tags":["bebida-bff-controller"],"operationId":"buscarBebidaPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DetalheBebidaDTO"}}}}}}},"/api/v1/bebidas/categorias":{"get":{"tags":["bebida-bff-controller"],"operationId":"listarCategorias","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/api/v1/bebidas/buscar":{"get":{"tags":["bebida-bff-controller"],"operationId":"buscarBebidas","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BebidaResumoDTO"}}}}}}}}},"components":{"schemas":{"UsuarioUpdateDTO":{"type":"object","properties":{"nome":{"type":"string"},"bio":{"type":"string"},"fotoAvatarUrl":{"type":"string"}}},"NovaExperienciaDTO":{"type":"object","properties":{"idBebida":{"type":"integer","format":"int64"},"nota":{"type":"number","format":"double","maximum":5,"minimum":0},"comentario":{"type":"string","maxLength":1000,"minLength":0},"visibilidade":{"type":"string"},"fotoPostUrl":{"type":"string"},"localizacao":{"type":"string"}},"required":["fotoPostUrl","idBebida","nota","visibilidade"]},"RegistroExperienciaDTO":{"type":"object","properties":{"idUsuario":{"type":"integer","format":"int64"},"itemId":{"type":"integer","format":"int64"},"nota":{"type":"number","format":"double"},"comentario":{"type":"string"},"data":{"type":"string"},"local":{"type":"string"},"fotoPostUrl":{"type":"string"}}},"NovoComentarioDTO":{"type":"object","properties":{"texto":{"type":"string","maxLength":500,"minLength":0}},"required":["texto"]},"ComentarioResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"texto":{"type":"string"},"tempoDecorrido":{"type":"string"},"autor":{"$ref":"#/components/schemas/UsuarioResumoDTO"}}},"UsuarioResumoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"fotoAvatarUrl":{"type":"string"}}},"UploadResponseDTO":{"type":"object","properties":{"url":{"type":"string"}}},"NovaBebidaDTO":{"type":"object","properties":{"nome":{"type":"string"},"fabricante":{"type":"string"},"categoria":{"type":"string"},"caracteristicas":{"type":"object","additionalProperties":{"type":"string"}}}},"BebidaResumoDTO":{"type":"object","properties":{"idBebida":{"type":"integer","format":"int64"},"nome":{"type":"string"},"categoria":{"type":"string"}}},"EstatisticasDTO":{"type":"object","properties":{"totalDegustacoes":{"type":"integer","format":"int32"},"notaMediaGlobal":{"type":"number","format":"double"},"seguindo":{"type":"integer","format":"int32"},"seguidores":{"type":"integer","format":"int32"}}},"PerfilDTO":{"type":"object","properties":{"seguindoPorMim":{"type":"boolean"},"usuario":{"$ref":"#/components/schemas/UsuarioPerfilDTO"},"estatisticas":{"$ref":"#/components/schemas/EstatisticasDTO"}}},"UsuarioPerfilDTO":{"type":"object","properties":{"idUsuario":{"type":"integer","format":"int64"},"nome":{"type":"string"},"bio":{"type":"string"},"fotoAvatarUrl":{"type":"string"}}},"PaginaBffDTOReciveUsuarioResumoDTO":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/UsuarioResumoDTO"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"NotificacaoResponseDTO":{"type":"object","properties":{"idNotificacao":{"type":"integer","format":"int64"},"tipo":{"type":"string"},"usuarioOrigem":{"$ref":"#/components/schemas/UsuarioOrigemDTO"},"mensagem":{"type":"string"},"tempoAtras":{"type":"string"},"lida":{"type":"boolean"}}},"PaginaBffDTOReciveNotificacaoResponseDTO":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/NotificacaoResponseDTO"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"UsuarioOrigemDTO":{"type":"object","properties":{"nome":{"type":"string"},"fotoAvatarUrl":{"type":"string"}}},"ContagemNotificacoesDTO":{"type":"object","properties":{"total":{"type":"integer","format":"int32"}}},"FeedResponseDTO":{"type":"object","properties":{"idPost":{"type":"integer","format":"int64"},"idUsuario":{"type":"integer","format":"int64"},"nomeAutor":{"type":"string"},"fotoAvatarUrl":{"type":"string"},"tempoDecorrido":{"type":"string"},"local":{"type":"string"},"fotoPostUrl":{"type":"string"},"idBebida":{"type":"integer","format":"int64"},"nomeBebida":{"type":"string"},"categoriaBebida":{"type":"string"},"nota":{"type":"number","format":"double"},"comentario":{"type":"string"},"curtidoPorMim":{"type":"boolean"},"totalCurtidas":{"type":"integer","format":"int32"},"totalComentarios":{"type":"integer","format":"int32"}}},"PaginaBffDTOReciveFeedResponseDTO":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/FeedResponseDTO"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"PaginaBffDTOReciveComentarioResponseDTO":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/ComentarioResponseDTO"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"DetalheBebidaDTO":{"type":"object","properties":{"idBebida":{"type":"integer","format":"int64"},"nome":{"type":"string"},"fabricante":{"type":"string"},"categoria":{"type":"string"},"notaMediaGlobal":{"type":"number","format":"double"},"caracteristicas":{"type":"object","additionalProperties":{"type":"string"}}}}}}}